re-jigged miner to use new algos at incorrect difficaulty
